ai-service/requirements.txt
psycopg[binary]==3.2.6
python-dotenv==1.0.1
fastapi==0.115.8
uvicorn==0.34.0
pytest==8.3.4
httpx==0.28.1
